Issue #62 - Modified the Report widget to set up the sql query in a separated method Issue #63 - Added query validation to the create/update process

+++ application/controllers/Reports.php
@@ -56,21 +56,24 @@ class Reports extends Easol_Controller {
         $model= new Easol_Report();
        //die(print_r($this->input->post('access[access]')));
         if($this->input->post('report') && $model->populateForm($this->input->post('report'))){
-                if($model->save()){
 
-                    $this->load->model('entities/easol/Easol_ReportAccess');
+                if ($this->form_validation->run() != FALSE) {
+                    if($model->save()){
 
-                    foreach($this->input->post('access[access]') as $access){
-                       $displayAccess = new Easol_ReportAccess();
-                        $displayAccess->ReportId = $model->ReportId;
-                        $displayAccess->RoleTypeId = $access;
-                        $displayAccess->save();
+                        $this->load->model('entities/easol/Easol_ReportAccess');
+
+                        foreach($this->input->post('access[access]') as $access){
+                           $displayAccess = new Easol_ReportAccess();
+                            $displayAccess->ReportId = $model->ReportId;
+                            $displayAccess->RoleTypeId = $access;
+                            $displayAccess->save();
+                        }
+                        $this->session->set_flashdata('message', 'New Report Added : '. $model->ReportName);
+                        $this->session->set_flashdata('type', 'success');
+
+                        return redirect(site_url("reports/index"));
+
                     }
-                    $this->session->set_flashdata('message', 'New Report Added : '. $model->ReportName);
-                    $this->session->set_flashdata('type', 'success');
-
-                    return redirect(site_url("reports/index"));
-
                 }
 
         }
@@ -98,44 +101,46 @@ class Reports extends Easol_Controller {
 
        // die(print_r($model));
         //die(print_r($this->input->post('access[access]')));
-        if($this->input->post('report') && $model->populateForm($this->input->post('report'))){
-            if($model->save()){
+        if($this->input->post('report') && $model->populateForm($this->input->post('report'))) {
 
-                $this->load->model('entities/easol/Easol_ReportAccess');
-/*
-                foreach($this->input->post('access[access]') as $access){
-                    $displayAccess = new Easol_ReportAccess();
-                    $displayAccess->ReportId = $model->ReportId;
-                    $displayAccess->RoleTypeId = $access;
-                    $displayAccess->save();
-                }
-*/
+            if ($this->form_validation->run() != FALSE) {
+                if($model->save()){
 
-                $aRoles=[];
-                foreach($model->getAccessTypes() as $role){
-                    if(!in_array($role->RoleTypeId,$this->input->post('access[access]'))){
-                        $this->db->delete('EASOL.ReportAccess', array('ReportId' => $model->ReportId,'RoleTypeId'=>$role->RoleTypeId));
+                    $this->load->model('entities/easol/Easol_ReportAccess');
+    /*
+                    foreach($this->input->post('access[access]') as $access){
+                        $displayAccess = new Easol_ReportAccess();
+                        $displayAccess->ReportId = $model->ReportId;
+                        $displayAccess->RoleTypeId = $access;
+                        $displayAccess->save();
                     }
-                    $aRoles[] = $role->RoleTypeId;
+    */
+
+                    $aRoles=[];
+                    foreach($model->getAccessTypes() as $role){
+                        if(!in_array($role->RoleTypeId,$this->input->post('access[access]'))){
+                            $this->db->delete('EASOL.ReportAccess', array('ReportId' => $model->ReportId,'RoleTypeId'=>$role->RoleTypeId));
+                        }
+                        $aRoles[] = $role->RoleTypeId;
+                    }
+
+                    foreach($this->input->post('access[access]') as $access){
+                        if(in_array($access,$aRoles))
+                            continue;
+                        $displayAccess = new Easol_ReportAccess();
+                        $displayAccess->ReportId = $model->ReportId;
+                        $displayAccess->RoleTypeId = $access;
+                        $displayAccess->save();
+                    }
+
+
+                    $this->session->set_flashdata('message', 'Report Updated Successfully : '. $model->ReportName);
+                    $this->session->set_flashdata('type', 'success');
+
+                    return  redirect(site_url("reports/edit/".$model->ReportId));
+
                 }
-
-                foreach($this->input->post('access[access]') as $access){
-                    if(in_array($access,$aRoles))
-                        continue;
-                    $displayAccess = new Easol_ReportAccess();
-                    $displayAccess->ReportId = $model->ReportId;
-                    $displayAccess->RoleTypeId = $access;
-                    $displayAccess->save();
-                }
-
-
-                $this->session->set_flashdata('message', 'Report Updated Successfully : '. $model->ReportName);
-                $this->session->set_flashdata('type', 'success');
-
-                return  redirect(site_url("reports/edit/".$model->ReportId));
-
             }
-
         }
         // $report->ReportName = "Report";
         /* echo $report->ReportName."sdd";
